// src/app/api/model/step-up-check-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StepUpChallengeType } from './step-up-challenge-type';


export interface StepUpCheckResponse { 
    challengeType?: StepUpChallengeType;
    required?: boolean;
}